backend/requirements.txt
# Core
fastapi>=0.115.5,<0.116
uvicorn[standard]>=0.32.0,<0.33
python-dotenv>=1.0.1,<1.1
aiofiles>=23.2.1,<24

# DB
motor>=3.6.1,<3.7
pymongo>=4.9.2,<4.10

# Vector store
chromadb>=0.5.11,<0.6
numpy>=1.26.4,<1.27

# Web search
tavily-python>=0.7.13,<0.8

# OpenAI
openai>=1.55.3,<1.56

# LangChain/Graph
langgraph>=0.2.39,<0.3
langchain>=0.3.14,<0.4
langchain-core>=0.3.29,<0.4
langchain-openai>=0.2.12,<0.3
langchain-text-splitters>=0.3.5,<0.4

# Email / Utils
sendgrid>=6.12.5,<6.13
pydub>=0.25.1,<0.26
tenacity>=9.0.0,<9.1
httpx>=0.27.2,<0.28

# Diarization (optional; needs PyTorch/CUDA set up)
pyannote.audio>=3.1,<3.2
